Effects.Concurrent.Thread

Description

Provides the MonadThread typeclass.

Since: 0.1

Thread Effect

class Monad m => MonadThread (m :: Type -> Type) where

Represents thread effects.

Since: 0.1

Methods

threadDelay :: Int -> m ()

Lifted threadDelay.

Since: 0.1

throwTo :: (Exception e, HasCallStack) => ThreadId -> e -> m ()

Lifted throwTo.

Since: 0.1

getNumCapabilities :: m Int

Lifted getNumCapabilities.

Since: 0.1

setNumCapabilities :: Int -> m ()

Lifted setNumCapabilities.

Since: 0.1

threadCapability :: ThreadId -> m (Int, Bool)

Lifted threadCapability.

Since: 0.1

myThreadId :: m ThreadId

Lifted myThreadId.

Since: 0.1

labelThread :: ThreadId -> String -> m ()

Lifted labelThread.

Since: 0.1

threadLabel :: ThreadId -> m (Maybe String)

Lifted threadLabel.

Since: 0.1

microsleep :: (HasCallStack, MonadThread m) => Natural -> m ()

threadDelay in terms of unbounded Natural rather than Int i.e. runs sleep in the current thread for the specified number of microseconds.

Since: 0.1

sleep :: (HasCallStack, MonadThread m) => Natural -> m ()

Runs sleep in the current thread for the specified number of seconds.

Since: 0.1

Reexports

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor

data ThreadId

A ThreadId is an abstract type representing a handle to a thread. ThreadId is an instance of Eq, Ord and Show, where the Ord instance implements an arbitrary total ordering over ThreadIds. The Show instance lets you convert an arbitrary-valued ThreadId to string form; showing a ThreadId value is occasionally useful when debugging or diagnosing the behaviour of a concurrent program.

Note: in GHC, if you have a ThreadId, you essentially have a pointer to the thread itself. This means the thread itself can't be garbage collected until you drop the ThreadId. This misfeature would be difficult to correct while continuing to support threadStatus.

MVar Effect

class Monad m => MonadMVar (m :: Type -> Type) where

Effect for MVar.

Since: 0.1

Methods

newEmptyMVar :: m (MVar a)

Lifted newEmptyMVar.

Since: 0.1

newMVar :: a -> m (MVar a)

Lifted newMVar.

Since: 0.1

newMVar' :: a -> m (MVar a)

Evaluates the input to newMVar to WHNF.

Since: 0.1

takeMVar :: MVar a -> m a

Lifted takeMVar.

Since: 0.1

takeMVar' :: MVar a -> m a

Evaluates the output of takeMVar to WHNF.

Since: 0.1

putMVar :: MVar a -> a -> m ()

Lifted putMVar.

Since: 0.1

putMVar' :: MVar a -> a -> m ()

Evaluates the input to putMVar to WHNF.

Since: 0.1

tryTakeMVar :: MVar a -> m (Maybe a)

Lifted tryTakeMVar.

Since: 0.1

tryTakeMVar' :: MVar a -> m (Maybe a)

Evaluates the output of tryTakeMVar to Nothing or Just a, where a is in WHNF.

Since: 0.1

tryPutMVar :: MVar a -> a -> m Bool

Lifted tryPutMVar.

Since: 0.1

tryPutMVar' :: MVar a -> a -> m Bool

Evaluates the input to tryPutMVar to WHNF.

Since: 0.1

isEmptyMVar :: MVar a -> m Bool

Lifted isEmptyMVar.

Since: 0.1

withMVar :: MVar a -> (a -> m b) -> m b

Lifted withMVar.

Since: 0.1

withMVar' :: MVar a -> (a -> m b) -> m b

Evaluates the input of the callback to withMVar to WHNF.

Since: 0.1

withMVarMasked :: MVar a -> (a -> m b) -> m b

Lifted withMVarMasked.

Since: 0.1

withMVarMasked' :: MVar a -> (a -> m b) -> m b

Evaluates the input of the callback to withMVarMasked to WHNF.

Since: 0.1

modifyMVar_ :: MVar a -> (a -> m a) -> m ()

Lifted modifyMVar_.

Since: 0.1

modifyMVar_' :: MVar a -> (a -> m a) -> m ()

Evaluates modifyMVar_'s modifier's input and output to WHNF.

Since: 0.1

modifyMVar :: MVar a -> (a -> m (a, b)) -> m b

Lifted modifyMVar.

Since: 0.1

modifyMVar' :: MVar a -> (a -> m (a, b)) -> m b

Evaluates modifyMVar's modifier's input and left-output to WHNF.

Since: 0.1

modifyMVarMasked_ :: MVar a -> (a -> m a) -> m ()

Lifted modifyMVarMasked_.

Since: 0.1

modifyMVarMasked_' :: MVar a -> (a -> m a) -> m ()

Evaluates modifyMVarMasked_'s modifier's input and output to WHNF.

Since: 0.1

modifyMVarMasked :: MVar a -> (a -> m (a, b)) -> m b

Lifted modifyMVarMasked.

Since: 0.1

modifyMVarMasked' :: MVar a -> (a -> m (a, b)) -> m b

Evaluates modifyMVarMasked's modifier's input and left-output to WHNF.

Since: 0.1

tryReadMVar :: MVar a -> m (Maybe a)

Lifted tryReadMVar.

Since: 0.1

tryReadMVar' :: MVar a -> m (Maybe a)

Evaluates the output of tryReadMVar to Nothing or Just a, where a is in WHNF.

Since: 0.1

mkWeakMVar :: MVar a -> m () -> m (Weak (MVar a))

Lifted mkWeakMVar.

Since: 0.1

Reexports

data MVar a

An MVar (pronounced "em-var") is a synchronising variable, used for communication between concurrent threads. It can be thought of as a box, which may be empty or full.

QSem Effect

class Monad m => MonadQSem (m :: Type -> Type) where

Effect for QSem semaphore.

Since: 0.1

Methods

newQSem :: Int -> m QSem

Lifted newQSem.

Since: 0.1

waitQSem :: QSem -> m ()

Lifted waitQSem.

Since: 0.1

signalQSem :: QSem -> m ()

Lifted signalQSem.

Since: 0.1

class Monad m => MonadQSemN (m :: Type -> Type) where

Effect for QSemN semaphore.

Since: 0.1

Methods

newQSemN :: Int -> m QSemN

Lifted newQSemN.

Since: 0.1

waitQSemN :: QSemN -> Int -> m ()

Lifted waitQSemN.

Since: 0.1

signalQSemN :: QSemN -> Int -> m ()

Lifted signalQSemN.

Since: 0.1

Reexports

data QSem

QSem is a quantity semaphore in which the resource is acquired and released in units of one. It provides guaranteed FIFO ordering for satisfying blocked waitQSem calls.

The pattern

bracket_ waitQSem signalQSem (...)

is safe; it never loses a unit of the resource.

data QSemN

QSemN is a quantity semaphore in which the resource is acquired and released in arbitrary amounts. It provides guaranteed FIFO ordering for satisfying blocked waitQSemN calls.

The pattern

bracket_ (waitQSemN n) (signalQSemN n) (...)

is safe; it never loses any of the resource.