Effects.FileSystem.Handle

Description

Typed file handles.

Since: 0.1

Typed handle

data Handle (p :: HandleMode)

Wrapper for file handles that includes an index for the mode.

Since: 0.1

unHandle :: forall (p :: HandleMode). Handle p -> Handle

Since: 0.1

unsafeHandle :: forall (p :: HandleMode). Handle -> Handle p

Makes a handle with any permissions. Useful for pseudo handles.

Since: 0.1

data HandleMode

Possible handle modes.

Since: 0.1

Constructors

HandleModeRead
Since: 0.1

HandleModeWrite
Since: 0.1

HandleModeReadWrite
Since: 0.1

IO handles

stdin :: Handle 'HandleModeReadWrite

Since: 0.1

stdout :: Handle 'HandleModeReadWrite

Since: 0.1

stderr :: Handle 'HandleModeReadWrite

Since: 0.1

Type families

type family CanRead (hm :: HandleMode) where ...

Since: 0.1

Equations

CanRead 'HandleModeRead = ()

CanRead 'HandleModeWrite = Unsatisfiable ('Text "HandleModeWrite does not have Read permission.")

CanRead 'HandleModeReadWrite = ()

type family CanWrite (hm :: HandleMode) where ...

Since: 0.1

Equations

CanWrite 'HandleModeRead = Unsatisfiable ('Text "HandleModeRead does not have Read permission.")

CanWrite 'HandleModeWrite = ()

CanWrite 'HandleModeReadWrite = ()

Aliases

type HandleIO = Handle

Alias for base's System.IO.Handle.

Since: 0.1

type HandleM (p :: HandleMode) = Handle p

Alias for typed Handle.

Since: 0.1

type HandleR = Handle 'HandleModeRead

Readable handle.

Since: 0.1

type HandleW = Handle 'HandleModeWrite

Writable handle.

Since: 0.1

type HandleRW = Handle 'HandleModeReadWrite

Readable and writable handle.

Since: 0.1

Locks

data LockedHandle (p :: HandleMode)

File handle with lock.

Since: 0.1

unLockedHandle :: forall (p :: HandleMode). LockedHandle p -> Handle p

Since: 0.1

unsafeLockedHandle :: forall (p :: HandleMode). Handle p -> LockedHandle p

Makes a LockedHandle without performing any locking.

Since: 0.1

liftLocked :: forall (p :: HandleMode) a. HasCallStack => (HasCallStack => Handle p -> a) -> LockedHandle p -> a

Lifts a function on handle to one on locked handles.

Since: 0.1

Aliases

type LockedHandleR = LockedHandle 'HandleModeRead

Readable handle.

Since: 0.1

type LockedHandleW = LockedHandle 'HandleModeWrite

Writable handle.

Since: 0.1

type LockedHandleRW = LockedHandle 'HandleModeReadWrite

Readable and writable handle.

Since: 0.1