Effects.FileSystem.HandleRW

Description

Provides the MonadHandleRW effect, for opening files with read and write abilities.

Since: 0.1

Effect

class (MonadHandleReader m, MonadHandleWriter m) => MonadHandleRW (m :: Type -> Type) where

Represents handle read and write effects.

Since: 0.1

Methods

openBinaryFile :: OsPath -> m (Handle 'HandleModeReadWrite)

Opens a file in ReadWrite mode.

Since: 0.1

withBinaryFile :: HasCallStack => OsPath -> (Handle 'HandleModeReadWrite -> m a) -> m a

Opens a file in ReadWrite mode.

Since: 0.1

data Handle (p :: HandleMode)

Wrapper for file handles that includes an index for the mode.

Since: 0.1

data HandleMode

Possible handle modes.

Since: 0.1

Constructors

HandleModeReadWrite
Since: 0.1

Reexports

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.