Effects.FileSystem.HandleReader

Description

Provides the MonadHandleReader effect.

Since: 0.1

Effect

class Monad m => MonadHandleReader (m :: Type -> Type) where

Represents handle reader effects.

Since: 0.1

Methods

openBinaryFile :: OsPath -> m (Handle 'HandleModeRead)

Opens a file for reading.

Since: 0.1

withBinaryFile :: HasCallStack => OsPath -> (Handle 'HandleModeRead -> m a) -> m a

Opens a file for reading.

Since: 0.1

hClose :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m ()

Lifted hClose.

Since: 0.1

hFlush :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m ()

Lifted hFlush.

Since: 0.1

hIsEOF :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hIsEOF.

Since: 0.1

hGetBuffering :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m BufferMode

Lifted hGetBuffering.

Since: 0.1

hIsOpen :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hIsOpen.

Since: 0.1

hIsClosed :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hIsClosed.

Since: 0.1

hIsReadable :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hIsReadable.

Since: 0.1

hIsWritable :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hIsWritable.

Since: 0.1

hIsSeekable :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hIsSeekable.

Since: 0.1

hIsTerminalDevice :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hIsTerminalDevice.

Since: 0.1

hGetEcho :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hGetEcho.

Since: 0.1

hWaitForInput :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> Int -> m Bool

Lifted hWaitForInput.

Since: 0.1

hReady :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Lifted hReady.

Since: 0.1

hGetChar :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Char

Lifted hGetChar.

Since: 0.1

hGetLine :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m ByteString

Lifted hGetLine.

Since: 0.1

hGetContents :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m ByteString

Lifted hGetContents.

Since: 0.1

hGet :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> Int -> m ByteString

Lifted hGet.

Since: 0.1

hGetSome :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> Int -> m ByteString

Lifted hGetSome.

Since: 0.1

hGetNonBlocking :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> Int -> m ByteString

Lifted hGetNonBlocking.

Since: 0.1

hLockRaw :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m ()

Attempts to shared lock a file, blocking or throwing an exception upon failure.

Since: 0.1

hTryLockRaw :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m Bool

Attempts to shared lock a file.

Since: 0.1

hUnlockRaw :: forall (p :: HandleMode). (CanRead p, HasCallStack) => Handle p -> m ()

Unlocks a locked file.

Since: 0.1

data Handle (p :: HandleMode)

Wrapper for file handles that includes an index for the mode.

Since: 0.1

data HandleMode

Possible handle modes.

Since: 0.1

Constructors

HandleModeRead
Since: 0.1

type family CanRead (hm :: HandleMode) where ...

Since: 0.1

Equations

CanRead 'HandleModeRead = ()

CanRead 'HandleModeWrite = Unsatisfiable ('Text "HandleModeWrite does not have Read permission.")

CanRead 'HandleModeReadWrite = ()

Locking

These functions bring some type-safety to file locking. Consider:

main :: (MonadHandleReader m) => m ()
main = withBinaryFile path $ \h -> do
  hLockRaw handle
  bs <- readBytes @HandleModeRead h
  hUnlockRaw handle
  print bs

readBytes :: (CanRead p, MonadHandleReader m) => Handle p -> m ByteString
readBytes handle = hGet handle 1024

In this example, we could remove all locking logic from main and everything would still compile. On the other hand:

main :: (MonadHandleReader m) => m ()
main = withBinaryFile path $ \h -> withLockedFile h $ \lh -> do
  bs <- readBytes @HandleModeRead lh
  print bs

readBytes :: (CanRead p, MonadHandleReader m) => LockedHandle p -> m ByteString
readBytes lockedHandle = liftLocked (\h -> hGet h 1024) lockedHandle

Removing withLockedFile would cause a compilation error, since readBytes requires a LockedHandle. The idea is to write most of the program's logic in terms of LockedHandle, using liftLocked to lift Handle functions.

data LockedHandle (p :: HandleMode)

File handle with lock.

Since: 0.1

liftLocked :: forall (p :: HandleMode) a. HasCallStack => (HasCallStack => Handle p -> a) -> LockedHandle p -> a

Lifts a function on handle to one on locked handles.

Since: 0.1

withLockedFile

Arguments

:: forall (p :: HandleMode) m a. (CanRead p, HasCallStack, MonadHandleReader m, MonadMask m)

=> Handle p
Handle to lock.

-> (LockedHandle p -> m a)
Callback with locked handle.

-> m a

Runs a computation with a shared locked file.

Since: 0.1

withTryLockedFile

Arguments

:: forall (p :: HandleMode) m a. (CanRead p, HasCallStack, MonadHandleReader m, MonadMask m)

=> Handle p
Handle to lock.

-> (LockedHandle p -> m a)
Handle callback.

-> m (Maybe a)

Like withSharedLockedFile, except the lock attempt does not block.

Since: 0.1

hLock

Arguments

:: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m)

=> Handle p
Handle to lock.

-> m (LockedHandle p)

Like hLockRaw, but returns a LockedHandle.

Since: 0.1

hTryLock

Arguments

:: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m)

=> Handle p
Handle to lock.

-> m (Maybe (LockedHandle p))

Like hTryLockRaw, but returns a LockedHandle if it succeeds.

Since: 0.1

hUnlock

Arguments

:: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m)

=> LockedHandle p
Handle to unlock.

-> m (Handle p)

Like hUnlockRaw, but returns the original handle.

Since: 0.1

Raw

withLockedFileRaw

Arguments

:: forall (p :: HandleMode) m a. (CanRead p, HasCallStack, MonadHandleReader m, MonadMask m)

=> Handle p
Handle to lock.

-> m a
Callback with locked handle.

-> m a

withLockedFile without LockedHandle.

Since: 0.1

withTryLockedFileRaw

Arguments

:: forall (p :: HandleMode) m a. (CanRead p, HasCallStack, MonadHandleReader m, MonadMask m)

=> Handle p
Handle to lock.

-> m a
Handle callback.

-> m (Maybe a)

withTryLockedFileRaw without LockedHandle.

Since: 0.1

UTF-8 Utils

GetLine

hGetLineUtf8 :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> m (Either UnicodeException Text)

hGetLine that attempts a UTF-8 conversion.

Since: 0.1

hGetLineUtf8Lenient :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> m Text

hGetLine that converts to UTF-8 in lenient mode.

Since: 0.1

hGetLineUtf8ThrowM :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m, MonadThrow m) => Handle p -> m Text

hGetLine that throws UnicodeException if UTF-8 conversion fails.

Since: 0.1

GetContents

hGetContentsUtf8 :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> m (Either UnicodeException Text)

hGetContents that attempts a UTF-8 conversion.

Since: 0.1

hGetContentsUtf8Lenient :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> m Text

hGetContents that converts to UTF-8 in lenient mode.

Since: 0.1

hGetContentsUtf8ThrowM :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m, MonadThrow m) => Handle p -> m Text

hGetContents that throws UnicodeException if UTF-8 conversion fails.

Since: 0.1

Get

hGetUtf8 :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> Int -> m (Either UnicodeException Text)

hGet that attempts a UTF-8 conversion.

Since: 0.1

hGetUtf8Lenient :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> Int -> m Text

hGet that converts to UTF-8 in lenient mode.

Since: 0.1

hGetUtf8ThrowM :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m, MonadThrow m) => Handle p -> Int -> m Text

hGet that throws UnicodeException if UTF-8 conversion fails.

Since: 0.1

GetSome

hGetSomeUtf8 :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> Int -> m (Either UnicodeException Text)

hGetSome that attempts a UTF-8 conversion.

Since: 0.1

hGetSomeUtf8Lenient :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> Int -> m Text

hGetSome that converts to UTF-8 in lenient mode.

Since: 0.1

hGetSomeUtf8ThrowM :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m, MonadThrow m) => Handle p -> Int -> m Text

hGetSome that throws UnicodeException if UTF-8 conversion fails.

Since: 0.1

GetNonBlocking

hGetNonBlockingUtf8 :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> Int -> m (Either UnicodeException Text)

hGetNonBlocking that attempts a UTF-8 conversion.

Since: 0.1

hGetNonBlockingUtf8Lenient :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m) => Handle p -> Int -> m Text

hGetNonBlocking that converts to UTF-8 in lenient mode.

Since: 0.1

hGetNonBlockingUtf8ThrowM :: forall (p :: HandleMode) m. (CanRead p, HasCallStack, MonadHandleReader m, MonadThrow m) => Handle p -> Int -> m Text

hGetNonBlocking that throws UnicodeException if UTF-8 conversion fails.

Since: 0.1

Reexports

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

data Text

A space efficient, packed, unboxed Unicode text type.

data UnicodeException

An exception type for representing Unicode encoding errors.