Effects.FileSystem.HandleWriter

Description

Provides the MonadFileWriter effect.

Since: 0.1

Effect

class Monad m => MonadHandleWriter (m :: Type -> Type) where

Represents handle writer effects.

Since: 0.1

Methods

openBinaryFile :: OsPath -> Bool -> m (Handle 'HandleModeWrite)

Opens a file for writing. If the boolean is true, opens in AppendMode. Otherwise opens in WriteMode.

Since: 0.1

withBinaryFile :: HasCallStack => OsPath -> Bool -> (Handle 'HandleModeWrite -> m a) -> m a

Opens a file for writing. If the boolean is true, opens in AppendMode. Otherwise opens in WriteMode.

Since: 0.1

hClose :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> m ()

Lifted hClose.

Since: 0.1

hFlush :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> m ()

Lifted hFlush.

Since: 0.1

hSetFileSize :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> Integer -> m ()

Lifted hSetFileSize.

Since: 0.1

hSetBuffering :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> BufferMode -> m ()

Lifted hSetBuffering.

Since: 0.1

hSeek :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> SeekMode -> Integer -> m ()

Lifted hSeek.

Since: 0.1

hTell :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> m Integer

Lifted hTell.

Since: 0.1

hSetEcho :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> Bool -> m ()

Lifted hSetEcho.

Since: 0.1

hPut :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> ByteString -> m ()

Lifted hPut.

Since: 0.1

hPutNonBlocking :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> ByteString -> m ByteString

Lifted hPutNonBlocking.

Since: 0.1

hLockRaw :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> m ()

Attempts to exclusively lock a file, blocking or throwing an exception upon failure.

Since: 0.1

hTryLockRaw :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> m Bool

Attempts to exclusively lock a file.

Since: 0.1

hUnlockRaw :: forall (p :: HandleMode). (CanWrite p, HasCallStack) => Handle p -> m ()

Unlocks a locked file.

Since: 0.1

data Handle (p :: HandleMode)

Wrapper for file handles that includes an index for the mode.

Since: 0.1

data HandleMode

Possible handle modes.

Since: 0.1

Constructors

HandleModeWrite
Since: 0.1

type family CanWrite (hm :: HandleMode) where ...

Since: 0.1

Equations

CanWrite 'HandleModeRead = Unsatisfiable ('Text "HandleModeRead does not have Read permission.")

CanWrite 'HandleModeWrite = ()

CanWrite 'HandleModeReadWrite = ()

Locking

These functions bring some type-safety to file locking. Consider:

main :: (MonadHandleWriter m) => m ()
main = withBinaryFile path False $ \h -> do
  hLockRaw handle
  bs <- writeBytes @HandleModeWrite h
  hUnlockRaw handle
  print bs

writeBytes :: (CanWrite p, MonadHandleWriter m) => Handle p -> m ()
writeBytes handle = hPut handle "some bytes"

In this example, we could remove all locking logic from main and everything would still compile. On the other hand:

main :: (MonadHandleWriter m) => m ()
main = withBinaryFile path False $ \h -> withLockedFile h $ \lh -> do
  bs <- writeBytes @HandleModeWrite lh
  print bs

writeBytes :: (CanWrite p, MonadHandleWriter m) => LockedHandle p -> m ByteString
writeBytes lockedHandle = liftLocked (\h -> hPut h "some bytes") lockedHandle

Removing withLockedFile would cause a compilation error, since writeBytes requires a LockedHandle. The idea is to write most of the program's logic in terms of LockedHandle, using liftLocked to lift Handle functions.

data LockedHandle (p :: HandleMode)

File handle with lock.

Since: 0.1

liftLocked :: forall (p :: HandleMode) a. HasCallStack => (HasCallStack => Handle p -> a) -> LockedHandle p -> a

Lifts a function on handle to one on locked handles.

Since: 0.1

withLockedFile

Arguments

:: forall (p :: HandleMode) m a. (CanWrite p, HasCallStack, MonadHandleWriter m, MonadMask m)

=> Handle p
Handle to lock.

-> (LockedHandle p -> m a)
Callback with locked handle.

-> m a

Runs a computation with an exclusively locked file.

Since: 0.1

withTryLockedFile

Arguments

:: forall (p :: HandleMode) m a. (CanWrite p, HasCallStack, MonadHandleWriter m, MonadMask m)

=> Handle p
Handle to lock.

-> (LockedHandle p -> m a)
Handle callback.

-> m (Maybe a)

Like withSharedLockedFile, except the lock attempt does not block.

Since: 0.1

hLock

Arguments

:: forall (p :: HandleMode) m. (CanWrite p, HasCallStack, MonadHandleWriter m)

=> Handle p
Handle to lock.

-> m (LockedHandle p)

Like hLockRaw, but returns a LockedHandle.

Since: 0.1

hTryLock

Arguments

:: forall (p :: HandleMode) m. (CanWrite p, HasCallStack, MonadHandleWriter m)

=> Handle p
Handle to lock.

-> m (Maybe (LockedHandle p))

Like hTryLockRaw, but returns a LockedHandle if it succeeds.

Since: 0.1

hUnlock

Arguments

:: forall (p :: HandleMode) m. (CanWrite p, HasCallStack, MonadHandleWriter m)

=> LockedHandle p
Handle to unlock.

-> m (Handle p)

Like hUnlockRaw, but returns the original handle.

Since: 0.1

Raw

withLockedFileRaw

Arguments

:: forall (p :: HandleMode) m a. (CanWrite p, HasCallStack, MonadHandleWriter m, MonadMask m)

=> Handle p
Handle to lock.

-> m a
Callback with locked handle.

-> m a

withLockedFile without LockedHandle.

Since: 0.1

withTryLockedFileRaw

Arguments

:: forall (p :: HandleMode) m a. (CanWrite p, HasCallStack, MonadHandleWriter m, MonadMask m)

=> Handle p
Handle to lock.

-> m a
Handle callback.

-> m (Maybe a)

withTryLockedFileRaw without LockedHandle.

Since: 0.1

UTF-8 Utils

hPutUtf8 :: forall (p :: HandleMode) m. (CanWrite p, HasCallStack, MonadHandleWriter m) => Handle p -> Text -> m ()

Writes the UTF-8 text to the handle.

Since: 0.1

hPutNonBlockingUtf8 :: forall (p :: HandleMode) m. (CanWrite p, HasCallStack, MonadHandleWriter m) => Handle p -> Text -> m ByteString

Writes UTF-8 text to handle, returning leftover bytes.

Since: 0.1

Misc

die :: (HasCallStack, MonadHandleWriter m, MonadThrow m) => String -> m a

Write given error message to stderr and terminate with exitFailure.

Since: 0.1

Reexports

data BufferMode

Three kinds of buffering are supported: line-buffering, block-buffering or no-buffering. These modes have the following effects. For output, items are written out, or flushed, from the internal buffer according to the buffer mode:

line-buffering: the entire output buffer is flushed whenever a newline is output, the buffer overflows, a hFlush is issued, or the handle is closed.

block-buffering: the entire buffer is written out whenever it overflows, a hFlush is issued, or the handle is closed.

no-buffering: output is written immediately, and never stored in the buffer.

An implementation is free to flush the buffer more frequently, but not less frequently, than specified above. The output buffer is emptied as soon as it has been written out.

Similarly, input occurs according to the buffer mode for the handle:

line-buffering: when the buffer for the handle is not empty, the next item is obtained from the buffer; otherwise, when the buffer is empty, characters up to and including the next newline character are read into the buffer. No characters are available until the newline character is available or the buffer is full.

block-buffering: when the buffer for the handle becomes empty, the next block of data is read into the buffer.

no-buffering: the next input item is read and returned. The hLookAhead operation implies that even a no-buffered handle may require a one-character buffer.

The default buffering mode when a handle is opened is implementation-dependent and may depend on the file system object which is attached to that handle. For most implementations, physical files will normally be block-buffered and terminals will normally be line-buffered.

Constructors

NoBuffering
buffering is disabled if possible.

LineBuffering
line-buffering should be enabled if possible.

BlockBuffering (Maybe Int)
block-buffering should be enabled if possible. The size of the buffer is n items if the argument is Just n and is otherwise implementation-dependent.

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

data IOMode

See openFile

Constructors

ReadMode

WriteMode

AppendMode

ReadWriteMode

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

data SeekMode

A mode that determines the effect of hSeek hdl mode i.

Constructors

AbsoluteSeek
the position of hdl is set to i.

RelativeSeek
the position of hdl is set to offset i from the current position.

SeekFromEnd
the position of hdl is set to offset i from the end of the file.

data Text

A space efficient, packed, unboxed Unicode text type.