Effects.FileSystem.PathWriter

Description

Provides the MonadPathWriter effect.

Since: 0.1

Effect

class Monad m => MonadPathWriter (m :: Type -> Type) where

Represents file-system writer effects.

Since: 0.1

Methods

createDirectory :: OsPath -> m ()

Lifted createDirectory.

Since: 0.1

createDirectoryIfMissing

Arguments

:: Bool
Create its parents too?

-> OsPath
The path to the directory you want to make

-> m ()

Lifted createDirectoryIfMissing.

Since: 0.1

removeDirectory :: OsPath -> m ()

Lifted removeDirectory.

Since: 0.1

removeDirectoryRecursive :: OsPath -> m ()

Lifted removeDirectoryRecursive.

Since: 0.1

removePathForcibly :: OsPath -> m ()

Lifted removePathForcibly.

Since: 0.1

renameDirectory :: OsPath -> OsPath -> m ()

Lifted renameDirectory.

Since: 0.1

setCurrentDirectory :: OsPath -> m ()

Lifted setCurrentDirectory.

Since: 0.1

withCurrentDirectory :: HasCallStack => OsPath -> m a -> m a

Lifted withCurrentDirectory.

Since: 0.1

removeFile :: OsPath -> m ()

Lifted removeFile.

Since: 0.1

renameFile :: OsPath -> OsPath -> m ()

Lifted renameFile.

Since: 0.1

renamePath

Arguments

:: OsPath
Old path

-> OsPath
New path

-> m ()

Lifted renamePath.

Since: 0.1

copyFile

Arguments

:: OsPath
Source filename

-> OsPath
Destination filename

-> m ()

Lifted copyFile.

Since: 0.1

copyFileWithMetadata

Arguments

:: OsPath
Source file

-> OsPath
Destination file

-> m ()

Lifted copyFileWithMetadata.

Since: 0.1

createFileLink

Arguments

:: OsPath
path to the target file

-> OsPath
path of the link to be created

-> m ()

Lifted createFileLink.

Since: 0.1

createDirectoryLink

Arguments

:: OsPath
path to the target directory

-> OsPath
path of the link to be created

-> m ()

Lifted createDirectoryLink.

Since: 0.1

removeDirectoryLink :: OsPath -> m ()

Lifted removeDirectoryLink.

Since: 0.1

setPermissions :: OsPath -> Permissions -> m ()

Lifted setPermissions.

Since: 0.1

copyPermissions :: OsPath -> OsPath -> m ()

Lifted copyPermissions.

Since: 0.1

setAccessTime :: OsPath -> UTCTime -> m ()

Lifted setAccessTime.

Since: 0.1

setModificationTime :: OsPath -> UTCTime -> m ()

Lifted setModificationTime.

Since: 0.1

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

Copying

Config

data CopyDirConfig

Directory copying config.

Since: 0.1

Constructors

MkCopyDirConfig

Fields

overwrite :: !Overwrite

Overwrite behavior.

Since: 0.1

targetName :: !TargetName

TargetName behavior.

Since: 0.1

data Overwrite

Determines file/directory overwrite behavior.

Since: 0.1

Constructors

OverwriteNone
No overwriting allowed.

Since: 0.1

OverwriteDirectories
Allow overwriting directories.

Since: 0.1

OverwriteAll
Allow overwriting the target directory and all subpaths.

Since: 0.1

data TargetName

Determines how to name the target.

Since: 0.1

Constructors

TargetNameSrc
Uses the src dir as the dest name i.e. dest/<src>.

Since: 0.1

TargetNameLiteral !OsPath
Uses the given literal as the dest name i.e. dest/<targetName>.

Since: 0.1

TargetNameDest
Uses dest itself as the target i.e. dest/ (top-level copy).

Since: 0.1

defaultCopyDirConfig :: CopyDirConfig

Default config for copying directories.

>>> defaultCopyDirConfig
MkCopyDirConfig {overwrite = OverwriteNone, destName = TargetNameSrc}

Since: 0.1

Functions

copyDirectoryRecursive

Arguments

:: (HasCallStack, MonadIORef m, MonadMask m, MonadPathReader m, MonadPathWriter m)

=> OsPath
Source

-> OsPath
Destination

-> m ()

copyDirectoryRecursiveConfig with defaultCopyDirConfig.

Since: 0.1

copyDirectoryRecursiveConfig

Arguments

:: (HasCallStack, MonadIORef m, MonadMask m, MonadPathReader m, MonadPathWriter m)

=> CopyDirConfig
Config

-> OsPath
Source

-> OsPath
Destination

-> m ()

copyDirectoryRecursiveConfig cfg src dest copies the src and its contents into dest e.g.

copyDirectoryRecursiveConfig cfg "path/to/foo" "path/to/bar"

will create path/to/bar/foo, path/to/bar/<target>, or copy foo's contents directly into bar, depending on the value of targetName.

The atomicity semantics are as follows:

OverwriteNone: If an error is encountered, we roll back the successful writes by deleting the entire dest/<target>.

OverwriteDirectories: If an error is encountered, we attempt to delete all successfully written paths/directories. Because these deletes are performed sequentially, we cannot guarantee all are removed before the process is interrupted.

OverwriteAll: Same as OverwriteDirectories, except paths that were overwritten are not restored. That is, if a path dest/<src>/p is overwritten and an error later encountered, p is not restored.

Throws:

IOException: if dest does not exist or one of:

OverwriteNone and dest/<src> exists.

OverwriteDirectories and some dest/<target>/p would be overwritten.

Since: 0.1

copySymbolicLink

Arguments

:: (HasCallStack, MonadCatch m, MonadPathReader m, MonadPathWriter m)

=> OsPath
Source

-> OsPath
Dest

-> m ()

Copies the symbolic link without traversing the link i.e. copy the link itself. Does not throw an exception if the target does exist. Throws an IOException if the path is not a symbolic link.

Windows: We have to distinguish between file and directory links (Posix makes no such distinction). If the target does not exist or is not considered a directory (e.g. it could also be a link), we fall back to creating a file link.

Since: 0.1

Optics

_OverwriteNone :: Prism' Overwrite ()

Since: 0.1

_OverwriteDirectories :: Prism' Overwrite ()

Since: 0.1

_OverwriteAll :: Prism' Overwrite ()

Since: 0.1

_TargetNameSrc :: Prism' TargetName ()

Since: 0.1

_TargetNameLiteral :: Prism' TargetName OsPath

Since: 0.1

_TargetNameDest :: Prism' TargetName ()

Since: 0.1

Removing

The removeXIfExists functions should be understood as helper combinators for the obvious doesXExist -> removeX pattern. They should not be understood as a total "delete arbitrary path if it exists" pattern.

For instance, doesDirectoryExist will return true if the target of a symbolic link is a directory, yet removeDirectory will throw an exception. Thus these functions should only be used when the type (file, dir, symlink) of a (possibly non-extant) path is known.

removeFileIfExists

Arguments

:: (HasCallStack, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m Bool

Calls removeFile if doesFileExist is True. Returns true iff the path is deleted.

Since: 0.1

removeFileIfExists_

Arguments

:: (HasCallStack, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m ()

Variant of removeFileIfExists that ignores the return value.

Since: 0.1

removeDirectoryIfExists

Arguments

:: (HasCallStack, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m Bool

Calls removeDirectory if doesDirectoryExist is True. Returns true iff the path is deleted.

Since: 0.1

removeDirectoryIfExists_

Arguments

:: (HasCallStack, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m ()

Variant of removeDirectoryIfExists that ignores the return value.

Since: 0.1

removeDirectoryRecursiveIfExists

Arguments

:: (HasCallStack, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m Bool

Calls removeDirectoryRecursive if doesDirectoryExist is True. Returns true iff the path is deleted.

Since: 0.1

removeDirectoryRecursiveIfExists_

Arguments

:: (HasCallStack, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m ()

Variant of removeDirectoryRecursiveIfExists that ignores the return value.

Since: 0.1

removePathForciblyIfExists

Arguments

:: (HasCallStack, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m Bool

Calls removePathForcibly if doesPathExist is True. Returns true iff the path is deleted.

Since: 0.1

removePathForciblyIfExists_

Arguments

:: (HasCallStack, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m ()

Variant of removePathForciblyIfExists that ignores the return value.

Since: 0.1

Symbolic Links

removeSymbolicLink

Arguments

:: (HasCallStack, MonadCatch m, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m ()

Removes a symbolic link. On Windows, attempts to distinguish file and directory links (Posix makes no distinction).

Since: 0.1

removeSymbolicLinkIfExists

Arguments

:: (HasCallStack, MonadCatch m, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m Bool

Calls removeSymbolicLink if doesSymbolicLinkExist is True. Returns true iff the path is deleted.

Since: 0.1

removeSymbolicLinkIfExists_

Arguments

:: (HasCallStack, MonadCatch m, MonadPathReader m, MonadPathWriter m)

=> OsPath
.

-> m ()

Variant of removeSymbolicLinkIfExists that ignores the return value.

Since: 0.1

Reexports

data IOException

Exceptions that occur in the IO monad. An IOException records a more specific error type, a descriptive string and maybe the handle that was used when the error was flagged.

data Permissions

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

Constructors

UTCTime

Fields

utctDay :: Day

the day

utctDayTime :: DiffTime

the time from midnight, 0 <= t < 86401s (because of leap-seconds)