Effects.HTTP.Client

Description

Provides dynamic network effects.

Since: 0.1

Effect

class Monad m => MonadNetwork (m :: Type -> Type) where

Dynamic effect for network effects.

Since: 0.1

Methods

newManager

Arguments

:: ManagerSettings
.

-> m Manager

Since: 0.1

newTlsManager

Arguments

:: m Manager
.

Since: 0.1

newTlsManagerWith

Arguments

:: ManagerSettings
.

-> m Manager

Since: 0.1

applyDigestAuth

Arguments

:: (HasCallStack, MonadThrow n)

=> ByteString
.

-> ByteString

-> Request

-> Manager

-> m (n Request)

Since: 0.1

withResponse

Arguments

:: HasCallStack

=> Request
.

-> Manager

-> (Response (BodyReader m) -> m a)

-> m a

Since: 0.1

withResponseHistory

Arguments

:: HasCallStack

=> Request
.

-> Manager

-> (HistoriedResponse (BodyReader m) -> m a)

-> m a

Since: 0.1

brRead

Arguments

:: BodyReader m
.

-> m ByteString

Since: 0.1

brReadSome

Arguments

:: BodyReader m
.

-> Int

-> m LazyByteString

Since: 0.1

brConsume

Arguments

:: BodyReader m
.

-> m [ByteString]

Since: 0.1

type BodyReader (m :: Type -> k) = m ByteString

Since: 0.1

Helpers

readResponseJson

Arguments

:: (FromJSON a, HasCallStack, MonadCatch m, MonadNetwork m)

=> String
String url.

-> Response (BodyReader m)
Response.

-> m a

Helper for reading a response, decoding JSON.

readResponseUtf8

Arguments

:: (HasCallStack, MonadCatch m, MonadNetwork m)

=> String
String url.

-> Response (BodyReader m)
Response.

-> m Text

Helper for reading a response, decoding to UTF-8.

Since: 0.1

readResponse

Arguments

:: forall a m. (HasCallStack, MonadCatch m)

=> String
String url.

-> Response (BodyReader m)
Response.

-> (BodyReader m -> m a)
Consumer.

-> m a

Helper for reading a response, checking for status 200 and exceptions thrown by the consumer.

Since: 0.1

Exceptions

data NetworkStatusE

Constructors

MkNetworkStatusE String Status

data NetworkReadBodyE

Constructors

MkNetworkReadBodyE String SomeException

data NetworkDecodeUtf8E

Constructors

MkNetworkDecodeUtf8E ByteString UnicodeException

data NetworkDecodeJsonE

Constructors

MkNetworkDecodeJsonE ByteString String

Re-exports

data ManagerSettings

Settings for a Manager. Please use the defaultManagerSettings function and then modify individual settings. For more information, see http://www.yesodweb.com/book/settings-types.

Since 0.1.0

data Manager

Keeps track of open connections for keep-alive.

If possible, you should share a single Manager between multiple threads and requests.

Since 0.1.0

data Request

All information on how to connect to a host and what should be sent in the HTTP request.

If you simply wish to download from a URL, see parseRequest.

The constructor for this data type is not exposed. Instead, you should use either the defaultRequest value, or parseRequest to construct from a URL, and then use the records below to make modifications. This approach allows http-client to add configuration options without breaking backwards compatibility.

For example, to construct a POST request, you could do something like:

initReq <- parseRequest "http://www.example.com/path"
let req = initReq
            { method = "POST"
            }

For more information, please see http://www.yesodweb.com/book/settings-types.

Since 0.1.0

data Response body

A simple representation of the HTTP response.

Since 0.1.0