Effects.IORef

Description

Provides the MonadIORef class.

Since: 0.1

Effect

class Monad m => MonadIORef (m :: Type -> Type) where

IORef effect.

Since: 0.1

Methods

newIORef :: HasCallStack => a -> m (IORef a)

Lifted newIORef.

Since: 0.1

newIORef' :: HasCallStack => a -> m (IORef a)

Evaluates a to WHNF then calls newIORef.

Since: 0.1

readIORef :: HasCallStack => IORef a -> m a

Lifted readIORef.

Since: 0.1

readIORef' :: HasCallStack => IORef a -> m a

Evaluates the result of readIORef to WHNF.

Since: 0.1

writeIORef :: HasCallStack => IORef a -> a -> m ()

Lifted writeIORef.

Since: 0.1

writeIORef' :: HasCallStack => IORef a -> a -> m ()

Evaluates a to WHNF before calling writeIORef.

Since: 0.1

atomicWriteIORef :: HasCallStack => IORef a -> a -> m ()

Lifted atomicWriteIORef.

Since: 0.1

atomicWriteIORef' :: HasCallStack => IORef a -> a -> m ()

Evaluates a to WHNF before calling atomicWriteIORef.

Since: 0.1

modifyIORef :: HasCallStack => IORef a -> (a -> a) -> m ()

Lifted modifyIORef'.

Since: 0.1

modifyIORef' :: HasCallStack => IORef a -> (a -> a) -> m ()

Lifted modifyIORef'.

Since: 0.1

atomicModifyIORef :: HasCallStack => IORef a -> (a -> (a, b)) -> m b

Lifted atomicModifyIORef.

Since: 0.1

atomicModifyIORef' :: HasCallStack => IORef a -> (a -> (a, b)) -> m b

Lifted atomicModifyIORef'.

Since: 0.1

Utils

atomicModifyIORef_ :: (HasCallStack, MonadIORef m) => IORef a -> (a -> a) -> m ()

Atomically apply a function to the contents of an IORef and return the old and new values.

Since: 0.1

atomicModifyIORef'_ :: (HasCallStack, MonadIORef m) => IORef a -> (a -> a) -> m ()

Atomically apply a function to the contents of an IORef and return the old and new values. The result of the function is forced.

Since: 0.1

Reexports

data IORef a

A mutable variable in the IO monad.

>>> import GHC.Internal.Data.IORef
>>> r <- newIORef 0
>>> readIORef r
0
>>> writeIORef r 1
>>> readIORef r
1
>>> atomicWriteIORef r 2
>>> readIORef r
2
>>> modifyIORef' r (+ 1)
>>> readIORef r
3
>>> atomicModifyIORef' r (\a -> (a + 1, ()))
>>> readIORef r
4

See also STRef and MVar.