Effects.Logger

Description

Provides logging effect and utilities..

Since: 0.1

Effect

class Monad m => MonadLogger (m :: Type -> Type) where

A Monad which has the ability to log messages in some manner.

Minimal complete definition

Nothing

Methods

monadLoggerLog :: ToLogStr msg => Loc -> LogSource -> LogLevel -> msg -> m ()

default monadLoggerLog :: forall (m' :: Type -> Type) (t :: (Type -> Type) -> Type -> Type) msg. (MonadLogger m', MonadTrans t, MonadLogger (t m'), ToLogStr msg, m ~ t m') => Loc -> LogSource -> LogLevel -> msg -> m ()

Levels

data LogLevel

Constructors

LevelDebug

LevelInfo

LevelWarn

LevelError

LevelOther Text

levelTrace :: LogLevel

Since: 0.1

levelFatal :: LogLevel

Since: 0.1

Logging functions

Levels

logTrace :: Q Exp

Since: 0.1

logDebug :: Q Exp

Generates a function that takes a Text and logs a LevelDebug message. Usage:

$(logDebug) "This is a debug log message"

logInfo :: Q Exp

See logDebug

logWarn :: Q Exp

See logDebug

logError :: Q Exp

See logDebug

logOther :: Text -> Q Exp

Generates a function that takes a Text and logs a LevelOther message. Usage:

$(logOther "My new level") "This is a log message"

logFatal :: Q Exp

Since: 0.1

Level checks

guardLevel

Arguments

:: Applicative f

=> LogLevel
The configured log level to check against.

-> LogLevel
The log level for this action.

-> f ()
The logging action to run if the level passes.

-> f ()

guardLevel configLvl lvl m runs m iff shouldLog configLvl lvl. This can be useful for writing a logging function e.g.

  -- logs msg to file iff configLogLevel <= lvl e.g.
  -- configLogLevel := LevelWarn
  -- lvl            := LevelError
  logMsg lvl msg = do
  configLogLevel <- getConfigLogLevel -- e.g. ReaderT Env
  guardLevel configLogLevel lvl $ do
    logToFile msg

Since: 0.1

shouldLog

Arguments

:: LogLevel
The configured log level to check against.

-> LogLevel
Level for this log

-> Bool
Whether we should log

shouldLog configLvl lvl returns true iff configLvl <= lvl. Uses LogLevel's built-in ordering with special cases for Trace (LevelOther Trace) and Fatal (LevelOther Fatal). The ad-hoc ordering is thus:

  LevelOther "Trace"
    < LevelDebug
    < LevelInfo
    < LevelWarn
    < LevelError
    < LevelOther "Fatal"
    < LevelOther "<any>"

In other words, LogLevel's usual Ord is respected, with the additional cases. Note that any other LevelOther "custom" sit at the the highest level and compare via Text's Ord, just like LogLevel's usual Ord.

Since: 0.1

Formatting

data LogFormatter

Formatter for logs.

Since: 0.1

Constructors

MkLogFormatter

Fields

locStrategy :: !LocStrategy

How to log the code location.

Since: 0.1

newline :: !Bool

If true, append a newline.

Since: 0.1

threadLabel :: !Bool

Whether to include the thread's label set by labelThread. Falls back to the thread's ThreadId when the label has not been set.

Since: 0.1

timezone :: !Bool

Whether to include the timezone in the timestamp.

Since: 0.1

defaultLogFormatter :: Loc -> LogFormatter

LogFormatter with:

locStrategy = LocPartial loc
newline = True
threadLabel = False
timezone = False

Since: 0.1

data LocStrategy

Determines how we log location data.

Since: 0.1

Constructors

LocPartial !Loc
Logs the location with filename, line, col.

Since: 0.1

LocStable !Loc
Logs the location with filename.

Since: 0.1

LocNone
No location logging.

Since: 0.1

formatLog

Arguments

:: (HasCallStack, MonadThread m, MonadTime m, ToLogStr msg)

=> LogFormatter
Formatter to use.

-> LogLevel
The level in which to log.

-> msg
Message.

-> m LogStr
Formatted LogStr.

Produces a formatted LogStr.

Example

-- [timestamp][thread_label][code_loc][level] msg
[2022-02-08 10:20:05][thread-label][filename:1:2][Warn] msg

Since: 0.1

Optics

LogLevels

_LevelTrace :: Prism' LogLevel ()

Since: 0.1

_LevelInfo :: Prism' LogLevel ()

Since: 0.1

_LevelDebug :: Prism' LogLevel ()

Since: 0.1

_LevelWarn :: Prism' LogLevel ()

Since: 0.1

_LevelError :: Prism' LogLevel ()

Since: 0.1

_LevelOther :: Prism' LogLevel Text

Since: 0.1

_LevelFatal :: Prism' LogLevel ()

Since: 0.1

Reexports

data LogStr

Log message builder. Use (<>) to append two LogStr in O(1).

data Loc

A location within a source file.