Effects.Optparse.Completer

Description

Provides additional Completers.

Since: 0.1

Bash completion

Aggregate completers

compgenCwdPathsCompleter :: Completer

Completer for cwd paths that first tries compgen via bashCompleterQuiet, then falls back to cwdPathsCompleter.

Since: 0.1

compgenCwdPathsSuffixCompleter :: String -> Completer

compgenCwdPathsCompleter that filters on the given suffix.

@since -.1

Compen

bashCompleterQuiet :: String -> Completer

Like bashCompleter, except any compgen stderrs are swallowed. This can be nicer when completion errors make the output messy.

Since: 0.1

Pure haskell

cwdPathsCompleter :: Completer

Completer based on paths in the current directory. Does not require external programs like compgen.

Since: 0.1

cwdPathsCompleterFilter :: (String -> Bool) -> Completer

cwdPathsCompleter that runs an additional filter.

Since: 0.1

Misc

requote :: String -> String

Strongly quote the string we pass to compgen.

We need to do this so bash doesn't expand out any ~ or other chars we want to complete on, or emit an end of line error when seeking the close to the quote.