Effects.System.Posix

Description

Provides the MonadPosix typeclass.

Since: 0.1

Effect

class Monad m => MonadPosix (m :: Type -> Type) where

Class for unix effects.

Since: 0.1

Methods

setFileMode :: PosixPath -> FileMode -> m ()

Since: 0.1

setFdMode :: Fd -> FileMode -> m ()

Since: 0.1

setFileCreationMask :: FileMode -> m FileMode

Since: 0.1

fileAccess :: PosixPath -> Bool -> Bool -> Bool -> m Bool

Since: 0.1

fileExist :: PosixPath -> m Bool

Since: 0.1

getFileStatus :: PosixPath -> m FileStatus

Since: 0.1

getFdStatus :: Fd -> m FileStatus

Since: 0.1

getSymbolicLinkStatus :: PosixPath -> m FileStatus

Since: 0.1

createNamedPipe :: PosixPath -> FileMode -> m ()

Since: 0.1

createDevice :: PosixPath -> FileMode -> DeviceID -> m ()

Since: 0.1

createLink :: PosixPath -> PosixPath -> m ()

Since: 0.1

removeLink :: PosixPath -> m ()

Since: 0.1

createSymbolicLink :: PosixPath -> PosixPath -> m ()

Since: 0.1

readSymbolicLink :: PosixPath -> m PosixPath

Since: 0.1

rename :: PosixPath -> PosixPath -> m ()

Since: 0.1

setOwnerAndGroup :: PosixPath -> UserID -> GroupID -> m ()

Since: 0.1

setFdOwnerAndGroup :: Fd -> UserID -> GroupID -> m ()

Since: 0.1

setSymbolicLinkOwnerAndGroup :: PosixPath -> UserID -> GroupID -> m ()

Since: 0.1

setFileTimes :: PosixPath -> EpochTime -> EpochTime -> m ()

Since: 0.1

touchFile :: PosixPath -> m ()

Since: 0.1

setFileSize :: PosixPath -> FileOffset -> m ()

Since: 0.1

setFdSize :: Fd -> FileOffset -> m ()

Since: 0.1

getPathVar :: PosixPath -> PathVar -> m Limit

Since: 0.1

getFdPathVar :: Fd -> PathVar -> m Limit

Since: 0.1

PathType

data PathType

Path type.

Since: fs-utils-0.1

Constructors

PathTypeFile
Since: fs-utils-0.1

PathTypeDirectory
Since: fs-utils-0.1

PathTypeSymbolicLink
Since: fs-utils-0.1

PathTypeOther
Since: fs-utils-0.1

Functions

displayPathType :: IsString a => PathType -> a

String representation of PathType.

Since: fs-utils-0.1

throwIfWrongPathType :: (HasCallStack, MonadCatch m, MonadPosix m) => String -> PathType -> PosixPath -> m ()

Throws IOException if the path does not exist or the expected path type does not match actual.

Since: 0.1

isPathType :: (HasCallStack, MonadPosix m) => PathType -> PosixPath -> m Bool

Checks that the path type matches the expectation. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1

getPathType :: (HasCallStack, MonadPosix m) => PosixPath -> m PathType

Returns the type for a given path without following symlinks. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1