Effects.System.PosixCompat.Files

Description

Provides the MonadPosixCompatFiles typeclass.

Since: 0.1

Effect

class Monad m => MonadPosixCompatFiles (m :: Type -> Type) where

Class for unix-compat System.PosixCompat.Files effects.

Since: 0.1

Methods

setFileMode :: FilePath -> FileMode -> m ()

Since: 0.1

setFdMode :: Fd -> FileMode -> m ()

Since: 0.1

setFileCreationMask :: FileMode -> m FileMode

Since: 0.1

fileAccess :: FilePath -> Bool -> Bool -> Bool -> m Bool

Since: 0.1

fileExist :: FilePath -> m Bool

Since: 0.1

getFileStatus :: FilePath -> m FileStatus

Since: 0.1

getFdStatus :: Fd -> m FileStatus

Since: 0.1

getSymbolicLinkStatus :: FilePath -> m FileStatus

Since: 0.1

createNamedPipe :: FilePath -> FileMode -> m ()

Since: 0.1

createDevice :: FilePath -> FileMode -> DeviceID -> m ()

Since: 0.1

createLink :: FilePath -> FilePath -> m ()

Since: 0.1

removeLink :: FilePath -> m ()

Since: 0.1

createSymbolicLink :: FilePath -> FilePath -> m ()

Since: 0.1

readSymbolicLink :: FilePath -> m FilePath

Since: 0.1

rename :: FilePath -> FilePath -> m ()

Since: 0.1

setOwnerAndGroup :: FilePath -> UserID -> GroupID -> m ()

Since: 0.1

setFdOwnerAndGroup :: Fd -> UserID -> GroupID -> m ()

Since: 0.1

setSymbolicLinkOwnerAndGroup :: FilePath -> UserID -> GroupID -> m ()

Since: 0.1

setFileTimes :: FilePath -> EpochTime -> EpochTime -> m ()

Since: 0.1

touchFile :: FilePath -> m ()

Since: 0.1

setFileSize :: FilePath -> FileOffset -> m ()

Since: 0.1

setFdSize :: Fd -> FileOffset -> m ()

Since: 0.1

getPathVar :: FilePath -> PathVar -> m Limit

Since: 0.1

getFdPathVar :: Fd -> PathVar -> m Limit

Since: 0.1

PathType

data PathType

Path type.

Since: fs-utils-0.1

Constructors

PathTypeFile
Since: fs-utils-0.1

PathTypeDirectory
Since: fs-utils-0.1

PathTypeSymbolicLink
Since: fs-utils-0.1

PathTypeOther
Since: fs-utils-0.1

Functions

displayPathType :: IsString a => PathType -> a

String representation of PathType.

Since: fs-utils-0.1

throwIfWrongPathType :: (HasCallStack, MonadCatch m, MonadPosixCompatFiles m) => String -> PathType -> FilePath -> m ()

Throws IOException if the path does not exist or the expected path type does not match actual.

Since: 0.1

isPathType :: (HasCallStack, MonadPosixCompatFiles m) => PathType -> FilePath -> m Bool

Checks that the path type matches the expectation. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1

getPathType :: (HasCallStack, MonadPosixCompatFiles m) => FilePath -> m PathType

Returns the type for a given path without following symlinks. Throws IOException if the path does not exist or the type cannot be detected.

Since: 0.1