Effects.System.Process

Description

Provides the MonadProcess typeclass for process effects.

Since: 0.1

Effect

class Monad m => MonadProcess (m :: Type -> Type) where

Effect for launching processes. API largely follows process.

Since: 0.1

Methods

createProcess

Arguments

:: CreateProcess
.

-> m (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)

Since: 0.1

createProcess_

Arguments

:: String
.

-> CreateProcess

-> m (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)

Since: 0.1

callProcess :: FilePath -> [String] -> m ()

Since: 0.1

callCommand :: String -> m ()

Since: 0.1

spawnProcess :: FilePath -> [String] -> m ProcessHandle

Since: 0.1

spawnCommand :: String -> m ProcessHandle

Since: 0.1

readCreateProcess :: CreateProcess -> String -> m String

Since: 0.1

readProcess

Arguments

:: FilePath
.

-> [String]

-> String

-> m String

Since: 0.1

readCreateProcessWithExitCode

Arguments

:: CreateProcess
.

-> String

-> m (ExitCode, String, String)

Since: 0.1

readProcessWithExitCode

Arguments

:: FilePath
.

-> [String]

-> String

-> m (ExitCode, String, String)

Since: 0.1

withCreateProcess

Arguments

:: HasCallStack

=> CreateProcess
.

-> (Maybe Handle -> Maybe Handle -> Maybe Handle -> ProcessHandle -> m a)

-> m a

Since: 0.1

cleanupProcess

Arguments

:: (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)
.

-> m ()

Since: 0.1

getPid :: ProcessHandle -> m (Maybe Pid)

Since: 0.1

getCurrentPid :: m Pid

Since: 0.1

waitForProcess :: ProcessHandle -> m ExitCode

Since: 0.1

getProcessExitCode :: ProcessHandle -> m (Maybe ExitCode)

Since: 0.1

terminateProcess :: ProcessHandle -> m ()

Since: 0.1

interruptProcessGroupOf :: ProcessHandle -> m ()

Since: 0.1

createPipe :: m (Handle, Handle)

Since: 0.1

createPipeFd :: m (FD, FD)

Since: 0.1

Running sub-processes

createProcess

Arguments

:: MonadProcess m

=> CreateProcess
.

-> m (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)

Since: 0.1

createProcess_

Arguments

:: MonadProcess m

=> String
.

-> CreateProcess

-> m (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)

Since: 0.1

shell :: String -> CreateProcess

Construct a CreateProcess record for passing to createProcess, representing a command to be passed to the shell.

proc :: FilePath -> [String] -> CreateProcess

Construct a CreateProcess record for passing to createProcess, representing a raw command with arguments.

See RawCommand for precise semantics of the specified FilePath.

data CreateProcess

Constructors

CreateProcess

Fields

cmdspec :: CmdSpec

Executable & arguments, or shell command. If cwd is Nothing, relative paths are resolved with respect to the current working directory. If cwd is provided, it is implementation-dependent whether relative paths are resolved with respect to cwd or the current working directory, so absolute paths should be used to ensure portability.

cwd :: Maybe FilePath

Optional path to the working directory for the new process

env :: Maybe [(String, String)]

Optional environment (otherwise inherit from the current process)

std_in :: StdStream

How to determine stdin

std_out :: StdStream

How to determine stdout

std_err :: StdStream

How to determine stderr

close_fds :: Bool

Close all file descriptors except stdin, stdout and stderr in the new process (on Windows, only works if std_in, std_out, and std_err are all Inherit). This implementation will call close on every fd from 3 to the maximum of open files, which can be slow for high maximum of open files. XXX verify what happens with fds in nodejs child processes

create_group :: Bool

Create a new process group. On JavaScript this also creates a new session.

delegate_ctlc :: Bool

Delegate control-C handling. Use this for interactive console processes to let them handle control-C themselves (see below for details).

Since: process-1.2.0.0

detach_console :: Bool

Use the windows DETACHED_PROCESS flag when creating the process; does nothing on other platforms.

Since: process-1.3.0.0

create_new_console :: Bool

Use the windows CREATE_NEW_CONSOLE flag when creating the process; does nothing on other platforms.

Default: False

Since: process-1.3.0.0

new_session :: Bool

Use posix setsid to start the new process in a new session; starts process in a new session on JavaScript; does nothing on other platforms.

Since: process-1.3.0.0

child_group :: Maybe GroupID

Use posix setgid to set child process's group id; works for JavaScript when system running nodejs is posix. does nothing on other platforms.

Default: Nothing

Since: process-1.4.0.0

child_user :: Maybe UserID

Use posix setuid to set child process's user id; works for JavaScript when system running nodejs is posix. does nothing on other platforms.

Default: Nothing

Since: process-1.4.0.0

use_process_jobs :: Bool

On Windows systems this flag indicates that we should wait for the entire process tree to finish before unblocking. On POSIX systems this flag is ignored. See $exec-on-windows for details.

Default: False

Since: process-1.5.0.0

data CmdSpec

Constructors

ShellCommand String
A command line to execute using the shell

RawCommand FilePath [String]
The name of an executable with a list of arguments

The FilePath argument names the executable, and is interpreted according to the platform's standard policy for searching for executables. Specifically:

on Unix systems the execvp(3) semantics is used, where if the executable filename does not contain a slash (/) then the PATH environment variable is searched for the executable.

on Windows systems the Win32 CreateProcess semantics is used. Briefly: if the filename does not contain a path, then the directory containing the parent executable is searched, followed by the current directory, then some standard locations, and finally the current PATH. An .exe extension is added if the filename does not already have an extension. For full details see the documentation for the Windows SearchPath API.

Windows does not have a mechanism for passing multiple arguments. When using RawCommand on Windows, the command line is serialised into a string, with arguments quoted separately. Command line parsing is up individual programs, so the default behaviour may not work for some programs. If you are not getting the desired results, construct the command line yourself and use ShellCommand.

data StdStream

Constructors

Inherit
Inherit Handle from parent

UseHandle Handle
Use the supplied Handle

CreatePipe
Create a new pipe. The returned Handle will use the default encoding and newline translation mode (just like Handles created by openFile).

NoStream
Close the stream's file descriptor without passing a Handle. On POSIX systems this may lead to strange behavior in the child process because attempting to read or write after the file has been closed throws an error. This should only be used with child processes that don't use the file descriptor at all. If you wish to ignore the child process's output you should either create a pipe and drain it manually or pass a Handle that writes to /dev/null.

data ProcessHandle

A handle to a process, which can be used to wait for termination of the process using waitForProcess.

None of the process-creation functions in this library wait for termination: they all return a ProcessHandle which may be used to wait for the process later.

On Windows a second wait method can be used to block for event completion. This requires two handles. A process job handle and a events handle to monitor.

Simpler functions for common tasks

callProcess :: MonadProcess m => FilePath -> [String] -> m ()

Since: 0.1

callCommand :: MonadProcess m => String -> m ()

Since: 0.1

spawnProcess :: MonadProcess m => FilePath -> [String] -> m ProcessHandle

Since: 0.1

spawnCommand :: MonadProcess m => String -> m ProcessHandle

Since: 0.1

readCreateProcess :: MonadProcess m => CreateProcess -> String -> m String

Since: 0.1

readProcess

Arguments

:: MonadProcess m

=> FilePath
.

-> [String]

-> String

-> m String

Since: 0.1

readCreateProcessWithExitCode

Arguments

:: MonadProcess m

=> CreateProcess
.

-> String

-> m (ExitCode, String, String)

Since: 0.1

readProcessWithExitCode

Arguments

:: MonadProcess m

=> FilePath
.

-> [String]

-> String

-> m (ExitCode, String, String)

Since: 0.1

withCreateProcess

Arguments

:: (MonadProcess m, HasCallStack)

=> CreateProcess
.

-> (Maybe Handle -> Maybe Handle -> Maybe Handle -> ProcessHandle -> m a)

-> m a

Since: 0.1

cleanupProcess

Arguments

:: MonadProcess m

=> (Maybe Handle, Maybe Handle, Maybe Handle, ProcessHandle)
.

-> m ()

Since: 0.1

Related utilities

showCommandForUser :: FilePath -> [String] -> String

Given a program p and arguments args, showCommandForUser p args returns a string suitable for pasting into /bin/sh (on Unix systems) or CMD.EXE (on Windows).

type Pid = CPid

The platform specific type for a process identifier.

This is always an integral type. Width and signedness are platform specific.

Since: process-1.6.3.0

getPid :: MonadProcess m => ProcessHandle -> m (Maybe Pid)

Since: 0.1

getCurrentPid :: MonadProcess m => m Pid

Since: 0.1

waitForProcess :: MonadProcess m => ProcessHandle -> m ExitCode

Since: 0.1

getProcessExitCode :: MonadProcess m => ProcessHandle -> m (Maybe ExitCode)

Since: 0.1

terminateProcess :: MonadProcess m => ProcessHandle -> m ()

Since: 0.1

interruptProcessGroupOf :: MonadProcess m => ProcessHandle -> m ()

Since: 0.1

createPipe :: MonadProcess m => m (Handle, Handle)

Since: 0.1

createPipeFd :: MonadProcess m => m (FD, FD)

Since: 0.1

Re-exports

data Handle

Haskell defines operations to read and write characters from and to files, represented by values of type Handle. Each value of this type is a handle: a record used by the Haskell run-time system to manage I/O with file system objects. A handle has at least the following properties:

whether it manages input or output or both;

whether it is open, closed or semi-closed;

whether the object is seekable;

whether buffering is disabled, or enabled on a line or block basis;

a buffer (whose length may be zero).

Most handles will also have a current I/O position indicating where the next input or output operation will occur. A handle is readable if it manages only input or both input and output; likewise, it is writable if it manages only output or both input and output. A handle is open when first allocated. Once it is closed it can no longer be used for either input or output, though an implementation cannot re-use its storage while references remain to it. Handles are in the Show and Eq classes. The string produced by showing a handle is system dependent; it should include enough information to identify the handle for debugging. A handle is equal according to == only to itself; no attempt is made to compare the internal state of different handles for equality.