Effects.Time

Description

Provides the MonadTime class.

Since: 0.1

Effect

class Monad m => MonadTime (m :: Type -> Type) where

Time effect.

Since: 0.1

Methods

getSystemZonedTime :: m ZonedTime

Returns the zoned system time.

Since: 0.1

getTimeZone :: UTCTime -> m TimeZone

Lifted getTimeZone.

Since: 0.1

utcToLocalZonedTime :: UTCTime -> m ZonedTime

Lifted utcToLocalZonedTime.

Since: 0.1

loadLocalTZ :: m TZ

Lifted loadLocalTZ.

Since: 0.1

getMonotonicTime :: m Double

Return monotonic time in seconds, since some unspecified starting point.

Since: 0.1

getSystemTime :: (HasCallStack, MonadTime m) => m LocalTime

Returns the local system time.

Since: 0.1

Timing

withTiming :: (HasCallStack, MonadTime m) => m a -> m (TimeSpec, a)

Runs an action, returning the elapsed time.

Since: 0.1

withTiming_ :: (HasCallStack, MonadTime m) => m a -> m TimeSpec

withTiming but ignores the result value.

Since: 0.1

TimeSpec

data TimeSpec

Structure for holding time data. Eq and Ord are defined in terms of an equivalence class e.g.

MkTimeSpec s n === MkTimeSpec 0 (s * 1_000_000_000 + n)

Since: 0.1

Constructors

MkTimeSpec

Fields

sec :: !Natural

Seconds.

Since: 0.1

nsec :: !Natural

Nanoseconds.

Since: 0.1

Creation

fromSeconds :: Double -> TimeSpec

Converts Double seconds to a TimeSpec.

Since: 0.1

fromNanoSeconds :: Natural -> TimeSpec

Converts Natural nanoseconds to a TimeSpec.

Since: 0.1

Elimination

toSeconds :: TimeSpec -> Double

Converts a TimeSpec to Double seconds.

Since: 0.1

toNanoSeconds :: TimeSpec -> Natural

Converts a TimeSpec into Natural nanoseconds.

Since: 0.1

Operations

diffTimeSpec :: TimeSpec -> TimeSpec -> TimeSpec

Returns the absolute difference of two TimeSpecs.

Since: 0.1

normalizeTimeSpec :: TimeSpec -> TimeSpec

Normalizes nanoseconds < 1 second.

Since: 0.1

Formatting

formatLocalTime :: LocalTime -> String

Formats the LocalTime to YYYY-MM-DD HH:MM:SS.

Since: 0.1

formatZonedTime :: ZonedTime -> String

Formats the ZonedTime to YYYY-MM-DD HH:MM:SS Z.

Since: 0.1

Parsing

parseLocalTime :: MonadFail f => String -> f LocalTime

Parses the LocalTime from YYYY-MM-DD HH:MM:SS.

Since: 0.1

parseZonedTime :: MonadFail f => String -> f ZonedTime

Parses the ZonedTime from YYYY-MM-DD HH:MM:SS Z.

Known Timezones

Expand

UTC

UT

GMT

EST

EDT

CST

CDT

MST

MDT

PST

PDT

+HHMM (e.g. +1300)

Since: 0.1

Misc

getSystemTimeString :: (HasCallStack, MonadTime m) => m String

Retrieves the formatted LocalTime.

Since: 0.1

getSystemZonedTimeString :: (HasCallStack, MonadTime m) => m String

Retrieves the formatted ZonedTime.

Since: 0.1

Reexports

Time

data LocalTime

A simple day and time aggregate, where the day is of the specified parameter, and the time is a TimeOfDay. Conversion of this (as local civil time) to UTC depends on the time zone. Conversion of this (as local mean time) to UT1 depends on the longitude.

Constructors

LocalTime

Fields

localDay :: Day

localTimeOfDay :: TimeOfDay

data TimeZone

A TimeZone is a whole number of minutes offset from UTC, together with a name and a "just for summer" flag.

data UTCTime

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

data ZonedTime

A local time together with a time zone.

There is no Eq instance for ZonedTime. If you want to compare local times, use zonedTimeToLocalTime. If you want to compare absolute times, use zonedTimeToUTC.

Constructors

ZonedTime

Fields

zonedTimeToLocalTime :: LocalTime

zonedTimeZone :: TimeZone

TZ

data TZ

Algebra

class ASemigroup s where

Defines an additive semigroup.

Since: algebra-simple-0.1

Methods

(.+.) :: s -> s -> s infixl 6

Since: algebra-simple-0.1

class ASemigroup m => AMonoid m where

Defines a monoid over an additive semigroup.

Since: algebra-simple-0.1

Methods

zero :: m

Should satisfy:

-- identity
x .+. zero = x = zero .+. x

since 0.1

class MSemigroup r => MSemiSpace m r | m -> r where

Defines a "multiplicative semi space" over an MSemigroup. This generalizes the notion of a Semimodule \(M\) over a Ring \(R\) in that we assume no additive structure on the space itself.

Since: algebra-simple-0.1

Methods

(.*) :: m -> r -> m infixl 7

Since: algebra-simple-0.1

class (MGroup k, MSemiSpace v k) => MSpace v k | v -> k where

Defines a "multiplicative space" over an MGroup. This generalizes the notion of a Semivectorspace \(V\) over a Semifield \(K\) in that we assume no additive structure on the space itself.

Since: algebra-simple-0.1

Methods

(.%) :: v -> k -> v infixl 7

Since: algebra-simple-0.1

class (AMonoid m, MSemiSpace m r, Semiring r) => Semimodule m r | m -> r

Defines a semimodule over a semiring. This generalizes the notion of a Module \(M\) over a Ring \(R\) such that:

\(M\) is an AMonoid, not an AGroup.

\(R\) is a Semiring, not a Ring.

Since: algebra-simple-0.1

class (MSpace v k, Semifield k, Semimodule v k) => SemivectorSpace v k | v -> k

Defines a semivector space over a semifield. This generalizes the notion of a VectorSpace \(V\) over a Field \(K\) such that:

\(V\) is an AMonoid, not an AGroup.

\(K\) is a Semifield, not a Field.

Since: algebra-simple-0.1

class Normed s

Defines a "norm" on a given structure. We have the following laws:

Positive definiteness: norm x > 0 for x /= 0, norm 0 = 0.

Subadditivity: norm (x + y) <= norm x + norm y.

Inversion : norm x = norm (-x).

Sign : sgn x = x / norm x.

Notice these laws require various algebraic assumptions, from a semigroup up to a group structure. It is expected that types that implement the relevant typeclasses will follow the respective laws.

We do not require any of these typeclasses as constraints as it is possible that we may want to abstract over "positive" types with a type that does not have an additive semigroup instance (consider multiplicative groups).

Since: algebra-simple-0.1

Minimal complete definition

norm, sgn

class LowerBounded a where

Names the lower limit of a type. Types that also have a Bounded instance should define lowerBound === minBound. This can be derived with the anyclass strategy.

Examples

Expand

>>> -- -XDeriveAnyClass, -XDerivingStrategies
>>> data Foo = Foo1 | Foo2 deriving stock (Bounded, Show) deriving anyclass LowerBounded
>>> lowerBound @Foo
Foo1

Since: bounds-0.1

Minimal complete definition

Nothing

Methods

lowerBound :: a

Since: bounds-0.1

default lowerBound :: Bounded a => a

class UpperBoundless a

Types that have no upper bound.

Since: bounds-0.1