Effects.Concurrent.STM

Description

Provides classes for STM.

Since: 0.1

Effect

data STM a

A monad supporting atomic memory transactions.

class Monad m => MonadAtomic (m :: Type -> Type) where

Effect for atomically lifting STM actions. Note that this classes is intended for "IO-like" monads not "STM-like" monads -- hence has no STM instance -- as the semantics for "STM-like" and "can lift STM atomically" are different.

Since: 0.1

Methods

atomically :: HasCallStack => STM a -> m a

Lifted atomically.

Since: 0.1

TVar

data TVar a

Shared memory locations that support atomic memory transactions.

Strict

newTVar' :: a -> STM (TVar a)

Evaluates the input to newTVar to WHNF.

Since: 0.1

readTVar' :: TVar a -> STM a

Evaluates the output from readTVar to WHNF.

Since: 0.1

writeTVar' :: TVar a -> a -> STM ()

Evaluates the input to writeTVar to WHNF.

Since: 0.1

modifyTVar' :: TVar a -> (a -> a) -> STM ()

Strict version of modifyTVar.

Since: stm-2.3

Atomic

newTVarA' :: (HasCallStack, MonadAtomic m) => a -> m (TVar a)

Atomic newTVar'.

Since: 0.1

readTVarA' :: (HasCallStack, MonadAtomic m) => TVar a -> m a

Atomic readTVarA'.

Since: 0.1

writeTVarA' :: (HasCallStack, MonadAtomic m) => TVar a -> a -> m ()

Atomic writeTVarA'.

Since: 0.1

modifyTVarA' :: (HasCallStack, MonadAtomic m) => TVar a -> (a -> a) -> m ()

Atomic modifyTVar'.

Since: 0.1

Lazy

newTVar :: a -> STM (TVar a)

Create a new TVar holding a value supplied

readTVar :: TVar a -> STM a

Return the current value stored in a TVar.

writeTVar :: TVar a -> a -> STM ()

Write the supplied value into a TVar.

modifyTVar :: TVar a -> (a -> a) -> STM ()

Mutate the contents of a TVar. N.B., this version is non-strict.

Since: stm-2.3

Atomic

newTVarA :: (HasCallStack, MonadAtomic m) => a -> m (TVar a)

Create a new TVar holding a value supplied and lifts the result via atomically.

Since: 0.1

readTVarA :: (HasCallStack, MonadAtomic m) => TVar a -> m a

Return the current value stored in a TVar and lifts the result via atomically.

Since: 0.1

writeTVarA :: (HasCallStack, MonadAtomic m) => TVar a -> a -> m ()

Write the supplied value into a TVar and lifts the action via atomically.

Since: 0.1

modifyTVarA :: (HasCallStack, MonadAtomic m) => TVar a -> (a -> a) -> m ()

Atomic modifyTVar.

Since: 0.1

TBQueue

data TBQueue a

TBQueue is an abstract type representing a bounded FIFO channel.

Since: stm-2.4

Strict

readTBQueue' :: TBQueue a -> STM a

Evaluates the output from readTBQueue to WHNF.

Since: 0.1

tryReadTBQueue' :: TBQueue a -> STM (Maybe a)

Evaluates the output from tryReadTBQueue to Nothing or Just a, where a is in WHNF.

Since: 0.1

writeTBQueue' :: TBQueue a -> a -> STM ()

Evaluates the input to writeTBQueue to WHNF.

Since: 0.1

flushTBQueue' :: TBQueue a -> STM [a]

Evaluates the output from flushTBQueue to WHNF.

Since: 0.1

Atomic

readTBQueueA' :: (HasCallStack, MonadAtomic m) => TBQueue a -> m a

Atomic readTBQueue'.

Since: 0.1

tryReadTBQueueA' :: (HasCallStack, MonadAtomic m) => TBQueue a -> m (Maybe a)

Atomic tryReadTBQueue'.

Since: 0.1

writeTBQueueA' :: (HasCallStack, MonadAtomic m) => TBQueue a -> a -> m ()

Atomic writeTBQueue'.

Since: 0.1

flushTBQueueA' :: (HasCallStack, MonadAtomic m) => TBQueue a -> m [a]

Atomic flushTBQueue'.

Since: 0.1

Lazy

newTBQueue

Arguments

:: Natural
maximum number of elements the queue can hold

-> STM (TBQueue a)

Builds and returns a new instance of TBQueue.

readTBQueue :: TBQueue a -> STM a

Read the next value from the TBQueue.

tryReadTBQueue :: TBQueue a -> STM (Maybe a)

A version of readTBQueue which does not retry. Instead it returns Nothing if no value is available.

writeTBQueue :: TBQueue a -> a -> STM ()

Write a value to a TBQueue; blocks if the queue is full.

flushTBQueue :: TBQueue a -> STM [a]

Efficiently read the entire contents of a TBQueue into a list. This function never retries.

Since: stm-2.4.5

Atomic

newTBQueueA

Arguments

:: (HasCallStack, MonadAtomic m)

=> Natural
maximum number of elements the queue can hold

-> m (TBQueue a)

Builds and returns a new instance of TBQueue, lifting via atomically.

Since: 0.1

readTBQueueA :: (HasCallStack, MonadAtomic m) => TBQueue a -> m a

Read the next value from the TBQueue, lifting via atomically.

Since: 0.1

tryReadTBQueueA :: (HasCallStack, MonadAtomic m) => TBQueue a -> m (Maybe a)

A version of readTBQueue which does not retry. Instead it returns Nothing if no value is available. Lifts via atomically.

Since: 0.1

writeTBQueueA :: (HasCallStack, MonadAtomic m) => TBQueue a -> a -> m ()

Write a value to a TBQueue; blocks if the queue is full. Lifts via atomically.

Since: 0.1

flushTBQueueA :: (HasCallStack, MonadAtomic m) => TBQueue a -> m [a]

Efficiently read the entire contents of a TBQueue into a list. This function never retries. Lifts via atomically.

Since: 0.1

Reexports

data Natural

Natural number

Invariant: numbers <= 0xffffffffffffffff use the NS constructor