Effects.FileSystem.HandleWriter

Description

Provides the MonadFileWriter effect.

Since: 0.1

Effect

class Monad m => MonadHandleWriter (m :: Type -> Type) where

Represents handle writer effects.

Since: 0.1

Methods

openBinaryFile :: OsPath -> IOMode -> m Handle

Lifted openBinaryFileIO.

Since: 0.1

withBinaryFile :: HasCallStack => OsPath -> IOMode -> (Handle -> m a) -> m a

Lifted withBinaryFileIO.

Since: 0.1

hClose :: Handle -> m ()

Lifted hClose.

Since: 0.1

hFlush :: Handle -> m ()

Lifted hFlush.

Since: 0.1

hSetFileSize :: Handle -> Integer -> m ()

Lifted hSetFileSize.

Since: 0.1

hSetBuffering :: Handle -> BufferMode -> m ()

Lifted hSetBuffering.

Since: 0.1

hSeek :: Handle -> SeekMode -> Integer -> m ()

Lifted hSeek.

Since: 0.1

hTell :: Handle -> m Integer

Lifted hTell.

Since: 0.1

hSetEcho :: Handle -> Bool -> m ()

Lifted hSetEcho.

Since: 0.1

hPut :: Handle -> ByteString -> m ()

Lifted hPut.

Since: 0.1

hPutNonBlocking :: Handle -> ByteString -> m ByteString

Lifted hPutNonBlocking.

Since: 0.1

type OsPath = OsString

Type representing filenames/pathnames.

This type doesn't add any guarantees over OsString.

UTF-8 Utils

hPutUtf8 :: (HasCallStack, MonadHandleWriter m) => Handle -> Text -> m ()

Writes the UTF-8 text to the handle.

Since: 0.1

hPutNonBlockingUtf8 :: (HasCallStack, MonadHandleWriter m) => Handle -> Text -> m ByteString

Writes UTF-8 text to handle, returning leftover bytes.

Since: 0.1

Misc

die :: (HasCallStack, MonadHandleWriter m, MonadThrow m) => String -> m a

Write given error message to stderr and terminate with exitFailure.

Since: 0.1

Reexports

data BufferMode

Three kinds of buffering are supported: line-buffering, block-buffering or no-buffering. These modes have the following effects. For output, items are written out, or flushed, from the internal buffer according to the buffer mode:

line-buffering: the entire output buffer is flushed whenever a newline is output, the buffer overflows, a hFlush is issued, or the handle is closed.

block-buffering: the entire buffer is written out whenever it overflows, a hFlush is issued, or the handle is closed.

no-buffering: output is written immediately, and never stored in the buffer.

An implementation is free to flush the buffer more frequently, but not less frequently, than specified above. The output buffer is emptied as soon as it has been written out.

Similarly, input occurs according to the buffer mode for the handle:

line-buffering: when the buffer for the handle is not empty, the next item is obtained from the buffer; otherwise, when the buffer is empty, characters up to and including the next newline character are read into the buffer. No characters are available until the newline character is available or the buffer is full.

block-buffering: when the buffer for the handle becomes empty, the next block of data is read into the buffer.

no-buffering: the next input item is read and returned. The hLookAhead operation implies that even a no-buffered handle may require a one-character buffer.

The default buffering mode when a handle is opened is implementation-dependent and may depend on the file system object which is attached to that handle. For most implementations, physical files will normally be block-buffered and terminals will normally be line-buffered.

Constructors

NoBuffering
buffering is disabled if possible.

LineBuffering
line-buffering should be enabled if possible.

BlockBuffering (Maybe Int)
block-buffering should be enabled if possible. The size of the buffer is n items if the argument is Just n and is otherwise implementation-dependent.

data ByteString

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

data IOMode

See openFile

Constructors

ReadMode

WriteMode

AppendMode

ReadWriteMode

data Handle

Haskell defines operations to read and write characters from and to files, represented by values of type Handle. Each value of this type is a handle: a record used by the Haskell run-time system to manage I/O with file system objects. A handle has at least the following properties:

whether it manages input or output or both;

whether it is open, closed or semi-closed;

whether the object is seekable;

whether buffering is disabled, or enabled on a line or block basis;

a buffer (whose length may be zero).

Most handles will also have a current I/O position indicating where the next input or output operation will occur. A handle is readable if it manages only input or both input and output; likewise, it is writable if it manages only output or both input and output. A handle is open when first allocated. Once it is closed it can no longer be used for either input or output, though an implementation cannot re-use its storage while references remain to it. Handles are in the Show and Eq classes. The string produced by showing a handle is system dependent; it should include enough information to identify the handle for debugging. A handle is equal according to == only to itself; no attempt is made to compare the internal state of different handles for equality.

data SeekMode

A mode that determines the effect of hSeek hdl mode i.

Constructors

AbsoluteSeek
the position of hdl is set to i.

RelativeSeek
the position of hdl is set to offset i from the current position.

SeekFromEnd
the position of hdl is set to offset i from the end of the file.

data Text

A space efficient, packed, unboxed Unicode text type.