Effects.Haskeline

Description

Provides the MonadHaskeline typeclass.

Since: 0.1

Class

class Monad m => MonadHaskeline (m :: Type -> Type) where

Allows haskeline effects.

Since: 0.1

Methods

haveTerminalUI :: m Bool

Lifted haveTerminalUI.

Since: 0.1

getInputLine :: String -> m (Maybe String)

Lifted getInputLine.

Since: 0.1

getInputLineWithInitial :: String -> (String, String) -> m (Maybe String)

Lifted getInputLineWithInitial.

Since: 0.1

getInputChar :: String -> m (Maybe Char)

Lifted getInputChar.

Since: 0.1

getPassword :: Maybe Char -> String -> m (Maybe String)

Lifted getPassword.

Since: 0.1

waitForAnyKey :: String -> m Bool

Lifted waitForAnyKey.

Since: 0.1

outputStr :: String -> m ()

Lifted outputStr.

Since: 0.1

outputStrLn :: String -> m ()

Lifted outputStrLn.

Since: 0.1

getExternalPrint :: m (String -> IO ())

Lifted getExternalPrint.

Since: 0.1

getHistory :: m History

Lifted getHistory.

Since: 0.1

putHistory :: History -> m ()

Lifted putHistory.

Since: 0.1

modifyHistory :: (History -> History) -> m ()

Lifted modifyHistory.

Since: 0.1

withInterrupt :: HasCallStack => m a -> m a

Lifted withInterrupt.

Since: 0.1

handleInterrupt :: HasCallStack => m a -> m a -> m a

Lifted handleInterrupt.

Since: 0.1

ReaderT

We provide two (overlapping) ReaderT instances: A standard instance i.e.

  instance (MonadHaskeline m) => MonadHaskeline (ReaderT e m)

And one in terms of the concrete (abstract) haskeline environment:

  instance MonadHaskeline (ReaderT (InputTEnv m) m)

The latter is the intended way to run the application in real code:

  run :: (MonadHaskeline m, ...) => m ()

  -- Uses ReaderT (InputTEnv m) instance
  main :: IO ()
  main = runInputTDefault runReaderT run

On the other hand, the former is useful for deriving e.g.

  newtype AppT e m a = MkAppT {unAppT :: ReaderT e m a }
    deriving (Applicative, Functor, Monad) via (ReaderT e m)

  deriving instance (MonadHaskeline m) => MonadHaskeline (AppT env m)

This allows typical usage with some custom type that picks up the instances automatically.

Runners

runInputTEnv :: (MonadIO m, MonadMask m) => (InputTEnv m -> m a) -> m a

runInputTWith with default haskeline settings.

Since: 0.1

runInputTEnvWith

Arguments

:: (InputT m a -> m a)
InputT Runner.

-> (InputTEnv m -> m a)
Action.

-> m a
IO Result.

Runs ReaderT InputTEnv in IO with InputT runner.

Since: 0.1

Haskeline Re-exports

Types

data InputT (m :: Type -> Type) a

A monad transformer which carries all of the state and settings relevant to a line-reading application.

data InputTEnv (m :: Type -> Type)

The abstract environment used by InputT, for ReaderT usage.

Since: haskeline-0.8.4.0

IO Runners

runInputT :: (MonadIO m, MonadMask m) => Settings m -> InputT m a -> m a

Run a line-reading application. This function should suffice for most applications.

This function is equivalent to runInputTBehavior defaultBehavior. It uses terminal-style interaction if stdin is connected to a terminal and has echoing enabled. Otherwise (e.g., if stdin is a pipe), it uses file-style interaction.

If it uses terminal-style interaction, Prefs will be read from the user's ~/.haskeline file (if present). If it uses file-style interaction, Prefs are not relevant and will not be read.

runInputTBehavior :: (MonadIO m, MonadMask m) => Behavior -> Settings m -> InputT m a -> m a

Run a line-reading application according to the given behavior.

If it uses terminal-style interaction, Prefs will be read from the user's ~/.haskeline file (if present). If it uses file-style interaction, Prefs are not relevant and will not be read.

runInputTBehaviorWithPrefs :: (MonadIO m, MonadMask m) => Behavior -> Prefs -> Settings m -> InputT m a -> m a

Run a line-reading application.

Config

defaultSettings :: forall (m :: Type -> Type). MonadIO m => Settings m

A useful default. In particular:

defaultSettings = Settings {
          complete = completeFilename,
          historyFile = Nothing,
          autoAddHistory = True
          }

defaultBehavior :: Behavior

Read input from stdin. Use terminal-style interaction if stdin is connected to a terminal and has echoing enabled. Otherwise (e.g., if stdin is a pipe), use file-style interaction.

This behavior should suffice for most applications.

defaultPrefs :: Prefs

The default preferences which may be overwritten in the .haskeline file.